Feature #7291 - PublicationInstanciator refactoring: - implementing ComponentInstanceDeletion into PublicationBmEJB and adding integration tests to validate it - removing unnecessary dependencies on notation library because it does not exist anymore!!! - implementing ComponentInstanceDeletion into DefaultRatingService and adding integration tests to validate it - re-activating all publication unit tests that was into test-waiting folder
Wiring into Admin services the call of delete method of ComponentInstanceDeletion implementations.
