+++ ejb-core/publication/src/integration-test/resources/META-INF/services/test-org.silverpeas.util.BeanContainer
@@ -1,0 +1,1 @@
+org.silverpeas.util.CDIContainer
