+++ ejb-core/publication/src/integration-test/resources/META-INF/test-MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Created-By: Silverpeas
+Dependencies: com.fasterxml.jackson.jaxrs.jackson-jaxrs-json-provider, org.apache.xerces
