+++ ejb-core/publication/src/integration-test/resources/arquillian.xml
@@ -1,0 +1,19 @@
+<arquillian xmlns="http://jboss.org/schema/arquillian"
+            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xsi:schemaLocation="
+        http://jboss.org/schema/arquillian
+        http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
+  <!-- uncomment it for have a glance at the war content
+  <engine>
+    <property name="deploymentExportPath">D:/DEVT/temp</property>
+  </engine>
+  -->
+  <container qualifier="wildfly-managed" default="true">
+    <configuration>
+      <!-- uncomment it for debugging integration test -->
+      <!--
+      <property name="javaVmArguments">-agentlib:jdwp=transport=dt_socket,address=5005,server=y,suspend=y</property>
+      -->
+    </configuration>
+  </container>
+</arquillian>
