+++ ejb-core/publication/src/integration-test/resources/com/stratelia/webactiv/util/publication/ejb/test-publication-dao-dataset.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright (C) 2000 - 2013 Silverpeas
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU Affero General Public License as
+    published by the Free Software Foundation, either version 3 of the
+    License, or (at your option) any later version.
+
+    As a special exception to the terms and conditions of version 3.0 of
+    the GPL, you may redistribute this Program in connection with Free/Libre
+    Open Source Software ("FLOSS") applications as described in Silverpeas's
+    FLOSS exception.  You should have received a copy of the text describing
+    the FLOSS exception, and it is also available here:
+    "http://www.silverpeas.org/docs/core/legal/floss_exception.html"
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU Affero General Public License for more details.
+
+    You should have received a copy of the GNU Affero General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+-->
+
+<dataset>
+  <sb_publication_publifather pubid="100" nodeid="110" instanceid="kmelia200" aliasuserid="0"
+                              aliasdate="2009/10/18" puborder="0" />
+  <sb_publication_publifather pubid="101" nodeid="110" instanceid="kmelia200" aliasuserid="0"
+                              aliasdate="2009/10/18" puborder="5" />
+  <sb_publication_publi pubid="100"  infoid="0"  pubname="Publication 1" pubdescription="Première publication de test"  pubcreationdate="2008/11/18"
+                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="100" pubimportance="1"
+                        pubversion="[NULL]"  pubkeywords="test"  pubcontent="Contenu de la publication 1"
+                        pubstatus="Valid" pubupdatedate="2009/11/18"
+                        instanceid="kmelia200" pubupdaterid="200" pubvalidatedate="2008/11/18"
+                        pubvalidatorid="300" pubbeginhour="00:00" pubendhour="23:59" pubauthor="Homer Simpson"
+                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
+  <sb_publication_publi pubid="101"  infoid="0"  pubname="Publication 2" pubdescription="2ème publication de test"  pubcreationdate="2008/11/18"
+                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="101" pubimportance="5"
+                        pubversion="[NULL]"  pubkeywords="test"  pubcontent="Contenu de la publication 2"
+                        pubstatus="Valid" pubupdatedate="2009/11/18"
+                        instanceid="kmelia200" pubupdaterid="200" pubvalidatedate="2008/11/18"
+                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
+                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
+</dataset>
