+++ ejb-core/publication/src/integration-test/resources/com/stratelia/webactiv/util/publication/ejb/testGetSelectByFatherPK_true_userId.txt
