+++ ejb-core/publication/src/integration-test/resources/com/stratelia/webactiv/util/publication/ejb/testGetSelectByNameAndNodeId.txt
