+++ ejb-core/publication/src/main/java/com/stratelia/webactiv/publication/PublicationInstanciator.java
@@ -40,6 +40,7 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 
+@Deprecated
 public class PublicationInstanciator extends SQLRequest {
 
   public PublicationInstanciator(String fullPathName) {
