--- ejb-core/publication/src/test-awaiting/resources/com/stratelia/webactiv/util/publication/ejb/test-last-publication-dao-dataset.xml
@@ -1,95 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2000 - 2013 Silverpeas
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    As a special exception to the terms and conditions of version 3.0 of
-    the GPL, you may redistribute this Program in connection with Free/Libre
-    Open Source Software ("FLOSS") applications as described in Silverpeas's
-    FLOSS exception.  You should have received a copy of the text describing
-    the FLOSS exception, and it is also available here:
-    "http://www.silverpeas.org/docs/core/legal/floss_exception.html"
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
--->
-
-<dataset>
-  <sb_publication_publifather pubid="100" nodeid="110" instanceid="kmelia100" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="0" />
-  <sb_publication_publifather pubid="101" nodeid="110" instanceid="kmelia100" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="5" />
-  <sb_publication_publifather pubid="200" nodeid="210" instanceid="kmelia200" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="0" />
-  <sb_publication_publifather pubid="201" nodeid="210" instanceid="kmelia200" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="5" />
-  <sb_publication_publifather pubid="202" nodeid="210" instanceid="kmelia200" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="5" />
-  <sb_publication_publifather pubid="301" nodeid="310" instanceid="kmelia300" aliasuserid="0"
-                              aliasdate="2009/10/18" puborder="5" />
-
-  <sb_publication_publi pubid="100"  infoid="0"  pubname="Publication 1" pubdescription="Publication Valide avec dates ok"  pubcreationdate="2008/11/18"
-                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="100" pubimportance="1"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 1"
-                        pubstatus="Valid" pubupdatedate="2010/10/25"
-                        instanceid="kmelia100" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="00:00" pubendhour="23:59" pubauthor="Homer Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-  <sb_publication_publi pubid="101"  infoid="0"  pubname="Publication 2" pubdescription="Publication 2 Valide avec dates ok"  pubcreationdate="2008/11/18"
-                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="101" pubimportance="5"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 2"
-                        pubstatus="Valid" pubupdatedate="2010/10/26"
-                        instanceid="kmelia100" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-  <sb_publication_publi pubid="103"  infoid="0"  pubname="Publication 3" pubdescription="Publication Valide avec dates not ok"  pubcreationdate="2008/11/18"
-                        pubbegindate="2009/10/18" pubenddate="2009/12/18"  pubcreatorid="101" pubimportance="5"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 2"
-                        pubstatus="Valid" pubupdatedate="2010/10/27"
-                        instanceid="kmelia100" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-
-
-  <sb_publication_publi pubid="200"  infoid="0"  pubname="Publication 4" pubdescription="Publication Valide avec dates ok"  pubcreationdate="2008/11/18"
-                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="100" pubimportance="1"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 1"
-                        pubstatus="Valid" pubupdatedate="2010/10/28"
-                        instanceid="kmelia200" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="00:00" pubendhour="23:59" pubauthor="Homer Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-  <sb_publication_publi pubid="201"  infoid="0"  pubname="Publication 2" pubdescription="Publication Draft avec dates ok"  pubcreationdate="2008/11/18"
-                        pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="101" pubimportance="5"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 2"
-                        pubstatus="Draft" pubupdatedate="2010/10/29"
-                        instanceid="kmelia200" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-  <sb_publication_publi pubid="202"  infoid="0"  pubname="Publication 2" pubdescription="Publication Valide avec updatedate trop vieille"
-                        pubcreationdate="2008/11/18" pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="100" pubimportance="1"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 2"
-                        pubstatus="Valid" pubupdatedate="2009/11/15"
-                        instanceid="kmelia200" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-
-  <sb_publication_publi pubid="301"  infoid="0"  pubname="Publication 2" pubdescription="Publication Valide avec dates ok dans kmelia hors scope"
-                        pubcreationdate="2008/11/18" pubbegindate="2009/10/18" pubenddate="2020/12/18"  pubcreatorid="101" pubimportance="5"
-                        pubversion=""  pubkeywords="test"  pubcontent="Contenu de la publication 2"
-                        pubstatus="Valid" pubupdatedate="2009/11/30"
-                        instanceid="kmelia300" pubupdaterid="200" pubvalidatedate="2008/11/18"
-                        pubvalidatorid="300" pubbeginhour="01:10" pubendhour="20:35" pubauthor="Bart Simpson"
-                        pubtargetvalidatorid="[NULL]" pubcloneid="-1" pubclonestatus="[NULL]" lang="fr" />
-
-</dataset>
